Reorgnise entrypoints into handlers module
